python/pyproject.toml
[project]
name = "pylinguis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ted Neward", email = "ted@tedneward.com" }
]
requires-python = ">=3.14"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
]

[project.scripts]
pylinguis = "pylinguis:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = ["src", ".venv/lib/python3.14/site-packages"]
log_cli = true
log_cli_level = "INFO"
